Adding Rasa and Weather Service

+++ chatbot/src/main/java/com/chatbot/websocket/responseMapperWeather/Weather.java
@@ -4,13 +4,13 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class Weather {
-    int temperature;
+    double temperature;
 
-    public int getTemperature() {
+    public double getTemperature() {
         return temperature;
     }
 
-    public void setTemperature(int temperature) {
+    public void setTemperature(double temperature) {
         this.temperature = temperature;
     }
 }
